replace Integer to Long for project filter

+++ src/main/java/eu/ec/doris/kohesio/services/FiltersGenerator.java
@@ -15,27 +15,27 @@ public class FiltersGenerator {
     @Autowired
     SimilarityService similarityService;
 
-        public String filterProject(String keywords,
-                                     String country,
-                                     String theme,
-                                     String fund,
-                                     String program,
-                                     String categoryOfIntervention,
-                                 String policyObjective,
-                                     Integer budgetBiggerThen,
-                                     Integer budgetSmallerThen,
-                                     Integer budgetEUBiggerThen,
-                                     Integer budgetEUSmallerThen,
-                                     String startDateBefore,
-                                     String startDateAfter,
-                                 String endDateBefore,
-                                 String endDateAfter,
-                                 String latitude,
-                                 String longitude,
-                                 String region,
-                                 String granularityRegion,
-                                 Integer limit,
-                                 Integer offset) throws IOException {
+    public String filterProject(String keywords,
+                                String country,
+                                String theme,
+                                String fund,
+                                String program,
+                                String categoryOfIntervention,
+                                String policyObjective,
+                                Long budgetBiggerThen,
+                                Long budgetSmallerThen,
+                                Long budgetEUBiggerThen,
+                                Long budgetEUSmallerThen,
+                                String startDateBefore,
+                                String startDateAfter,
+                                String endDateBefore,
+                                String endDateAfter,
+                                String latitude,
+                                String longitude,
+                                String region,
+                                String granularityRegion,
+                                Integer limit,
+                                Integer offset) throws IOException {
         String search = "";
         if (keywords != null) {
 //            if (!keywords.contains("AND") && !keywords.contains("OR") && !keywords.contains("NOT")) {
@@ -187,38 +187,39 @@ public class FiltersGenerator {
                 "   ?s0 <https://linkedopendata.eu/prop/direct/P35> <https://linkedopendata.eu/entity/Q9934> . ";
         return search;
     }
-    public HashMap<String,List<String>> getCountriesCodeMapping(){
+
+    public HashMap<String, List<String>> getCountriesCodeMapping() {
         HashMap<String, List<String>> mapping = new HashMap<>();
 //        EU_LANGUAGES = ['bg', 'cs', 'da', 'de', 'el', 'en', 'es', 'et', 'fi', 'fr', 'ga', 'hr',
 //                'hu', 'it', 'lt', 'lv', 'mt', 'nl', 'pl', 'pt', 'ro', 'sk', 'sl', 'sv']
 
         mapping.put("<https://linkedopendata.eu/entity/Q14>", Arrays.asList("lt")); // lithuania
-        mapping.put("<https://linkedopendata.eu/entity/Q10>",Arrays.asList("fi")); // finland
-        mapping.put("<https://linkedopendata.eu/entity/Q27>",Arrays.asList("sl")); // slovenia
-        mapping.put("<https://linkedopendata.eu/entity/Q18>",Arrays.asList("pt")); // portugal
-        mapping.put("<https://linkedopendata.eu/entity/Q23>",Arrays.asList("et")); // estonia
-        mapping.put("<https://linkedopendata.eu/entity/Q9>",Arrays.asList("fr")); // luxembourg
-        mapping.put("<https://linkedopendata.eu/entity/Q29>",Arrays.asList("bg")); // bulgaria
-        mapping.put("<https://linkedopendata.eu/entity/Q28>",Arrays.asList("ro")); // romania
-        mapping.put("<https://linkedopendata.eu/entity/Q3>",Arrays.asList("hu")); // hungary
-        mapping.put("<https://linkedopendata.eu/entity/Q19>",Arrays.asList("nl")); // netherlands
-        mapping.put("<https://linkedopendata.eu/entity/Q26>",Arrays.asList("sk")); // slovakia
-        mapping.put("<https://linkedopendata.eu/entity/Q16>",Arrays.asList("de")); // austria
-        mapping.put("<https://linkedopendata.eu/entity/Q24>",Arrays.asList("lv")); // latvia
-        mapping.put("<https://linkedopendata.eu/entity/Q12>",Arrays.asList("da")); // denmark
-        mapping.put("<https://linkedopendata.eu/entity/Q30>",Arrays.asList("hr")); // croatia
-        mapping.put("<https://linkedopendata.eu/entity/Q11>",Arrays.asList("sv")); // sweden
-        mapping.put("<https://linkedopendata.eu/entity/Q7>",Arrays.asList("es")); // spain
-        mapping.put("<https://linkedopendata.eu/entity/Q22>",Arrays.asList("de")); // germany
-        mapping.put("<https://linkedopendata.eu/entity/Q31>",Arrays.asList("el")); // cyprus
-        mapping.put("<https://linkedopendata.eu/entity/Q17>",Arrays.asList("el")); // greece
-        mapping.put("<https://linkedopendata.eu/entity/Q25>",Arrays.asList("cs")); // czech republic
-        mapping.put("<https://linkedopendata.eu/entity/Q13>",Arrays.asList("pl")); // poland
-        mapping.put("<https://linkedopendata.eu/entity/Q15>",Arrays.asList("it")); // italy
-        mapping.put("<https://linkedopendata.eu/entity/Q20>",Arrays.asList("fr")); // france
-        mapping.put("<https://linkedopendata.eu/entity/Q2>",Arrays.asList("en")); // ireland
-        mapping.put("<https://linkedopendata.eu/entity/Q32>",Arrays.asList("en")); // malta
-        mapping.put("<https://linkedopendata.eu/entity/Q8>",Arrays.asList("fr","nl","de")); // belgium
+        mapping.put("<https://linkedopendata.eu/entity/Q10>", Arrays.asList("fi")); // finland
+        mapping.put("<https://linkedopendata.eu/entity/Q27>", Arrays.asList("sl")); // slovenia
+        mapping.put("<https://linkedopendata.eu/entity/Q18>", Arrays.asList("pt")); // portugal
+        mapping.put("<https://linkedopendata.eu/entity/Q23>", Arrays.asList("et")); // estonia
+        mapping.put("<https://linkedopendata.eu/entity/Q9>", Arrays.asList("fr")); // luxembourg
+        mapping.put("<https://linkedopendata.eu/entity/Q29>", Arrays.asList("bg")); // bulgaria
+        mapping.put("<https://linkedopendata.eu/entity/Q28>", Arrays.asList("ro")); // romania
+        mapping.put("<https://linkedopendata.eu/entity/Q3>", Arrays.asList("hu")); // hungary
+        mapping.put("<https://linkedopendata.eu/entity/Q19>", Arrays.asList("nl")); // netherlands
+        mapping.put("<https://linkedopendata.eu/entity/Q26>", Arrays.asList("sk")); // slovakia
+        mapping.put("<https://linkedopendata.eu/entity/Q16>", Arrays.asList("de")); // austria
+        mapping.put("<https://linkedopendata.eu/entity/Q24>", Arrays.asList("lv")); // latvia
+        mapping.put("<https://linkedopendata.eu/entity/Q12>", Arrays.asList("da")); // denmark
+        mapping.put("<https://linkedopendata.eu/entity/Q30>", Arrays.asList("hr")); // croatia
+        mapping.put("<https://linkedopendata.eu/entity/Q11>", Arrays.asList("sv")); // sweden
+        mapping.put("<https://linkedopendata.eu/entity/Q7>", Arrays.asList("es")); // spain
+        mapping.put("<https://linkedopendata.eu/entity/Q22>", Arrays.asList("de")); // germany
+        mapping.put("<https://linkedopendata.eu/entity/Q31>", Arrays.asList("el")); // cyprus
+        mapping.put("<https://linkedopendata.eu/entity/Q17>", Arrays.asList("el")); // greece
+        mapping.put("<https://linkedopendata.eu/entity/Q25>", Arrays.asList("cs")); // czech republic
+        mapping.put("<https://linkedopendata.eu/entity/Q13>", Arrays.asList("pl")); // poland
+        mapping.put("<https://linkedopendata.eu/entity/Q15>", Arrays.asList("it")); // italy
+        mapping.put("<https://linkedopendata.eu/entity/Q20>", Arrays.asList("fr")); // france
+        mapping.put("<https://linkedopendata.eu/entity/Q2>", Arrays.asList("en")); // ireland
+        mapping.put("<https://linkedopendata.eu/entity/Q32>", Arrays.asList("en")); // malta
+        mapping.put("<https://linkedopendata.eu/entity/Q8>", Arrays.asList("fr", "nl", "de")); // belgium
 
         return mapping;
     }
